Added make_hrr_from_string.
Instead of generating an HRR and then derferencing with the appropriate
vector, as make_hrr_noise_from_string does, this function just returns
the HRR vector itself, with no dereferencing. As a result, the input
string does not need to have a '!', and the returned function does not
need to take an input vector.

Also, factored out some of the code common between the two functions
into helper functions.