Added order. Stores index, data by order added.
The idea is that by storing this order, you can just add everything in
your simulation to the bootstrapper and have a fairly complete story of
what went on.

Later, might want to have something that does some introspection and
figures out which class or function added that item. Just a nice extra
peice of info.