Added filestrap decorator.
Easy persistent memoization using Bootstrap object.

Currently unsure about the memoization arguments. Should the user,
at decoration time, specify all the arguments that could possibly
be memoized, or all the arguments that ARE memozied? That is, when the
function is called, do we require that all memoized variables be
provided to the function? The alternative is to just memoize based on
presented arguments that have also been designated as memoizable.